// src/app/Types.ts
// src/types.ts

export interface Category {
    text: string;
    pic: string;
    courses: string;
    color: string;
  }
  
  export interface Course {
    label: string;
    price: string;
    title: string;
    picurl: string;
    stars: string;
    students: string;
    colors: string;
    date: string;
  }
  
  export interface Instructor {
    name: string;
    pic: string;
    title: string;
    raiting: string;
    students: string;
  }
  